Minor cleanup

+++ lib-ws28xxdmx/include/ws28xxstripeparams.h
@@ -55,7 +55,7 @@ public:
 	static const char *GetLedTypeString(TWS28XXType);
 
 private:
-	bool isMaskSet(uint16_t) const;
+	bool isMaskSet(uint32_t) const;
 
 public:
     static void staticCallbackFunction(void *p, const char *s);
